Decide CC brand by lambda, not regex
Previously, we were using pure regexes to determine a credit card brand
based on its BIN. The problem is that you can't reliably determine
BINs based on regexes: Maestro overlaps with MasterCard, Electron just
has a pile of buckets, and so on.

As a first step, remove any reliance on raw regexes, and instead switch
to using lambdas. For now, this is functionally equivalent, but will
allow more precise brand determination in a follow-up commit.

Unit: 3927 tests, 68196 assertions, 0 failures, 0 errors, 0 pendings, 2 omissions, 0 notifications
100% passed